Initial changes for extension channel
