+++ src/Worker.Extensions.DurableTask/Worker.Extensions.DurableTask.csproj
@@ -36,10 +36,11 @@
   </PropertyGroup>
   
   <ItemGroup>
-    <PackageReference Include="Microsoft.Azure.Functions.Worker.Core" Version="1.8.0" />
-    <PackageReference Include="Microsoft.Azure.Functions.Worker.Extensions.Abstractions" Version="1.1.0" />
-    <PackageReference Include="Microsoft.DurableTask.Client.Grpc" Version="1.0.2" />
-    <PackageReference Include="Microsoft.DurableTask.Worker.Grpc" Version="1.0.2" />
+    <PackageReference Include="Microsoft.Azure.Functions.Worker.Core" Version="1.18.0" />
+    <PackageReference Include="Microsoft.Azure.Functions.Worker.Extensions.Abstractions" Version="1.3.0" />
+    <PackageReference Include="Microsoft.DurableTask.Client.Grpc" Version="1.2.2" />
+    <PackageReference Include="Microsoft.DurableTask.Worker.Grpc" Version="1.2.2" />
+    <PackageReference Include="Microsoft.Azure.Functions.Worker.Extensions.Rpc" Version="1.0.0" />
   </ItemGroup>
 
   <ItemGroup>
